Adding authentication, split the management into plots, major bank refactor
