+++ .dockerignore
@@ -6,5 +6,4 @@ __pycache__/
 .env
 .idea
 .venv
-reports
 .DS_Store
