+++ .gitignore
@@ -8,3 +8,4 @@ __pycache__/
 reports
 .idea
 .DS_Store
+.vscode
